+++ threadfunctions.h
@@ -1,10 +1,8 @@
 #ifndef THREADFUNCTIONS_H
 #define THREADFUNCTIONS_H
 
-#include <cstddef>
-
 unsigned GetHardwareConcurrency();
 
-double SingleThreadCalculation(size_t i_thread);
+double SingleThreadCalculation(unsigned i_thread);
 
 #endif // THREADFUNCTIONS_H
